Move HumanDate out of signoff/
