+++ src/components/signoff/Review.tsx
@@ -1,7 +1,7 @@
 import AdminLink from "../AdminLink";
 import { Comment } from "./Comment";
-import HumanDate from "./HumanDate";
 import { ProgressStep } from "./ProgressBar";
+import HumanDate from "@src/components/HumanDate";
 import { useSimpleReview } from "@src/hooks/preferences";
 import type {
   ChangesList,
